requirements.txt
langchain
langchain-openai
langchain-google-genai
python-dotenv
openai
gradio

pydantic
langchain-google
langgraph
wikipedia
langchain-community
faiss-cpu